Add partial features for year updates & runup phase
So that we don't have to edit the copy to edit the year every time.

And, so that we can have one template and code path that twigs
on the current date to change the rendered text at the right times.

`page.tpl` still needs a lot of work, including adding some `RUNUP`
checks. Sometimes this will be a separate `elif RUNUP`, other times
we'll need an `if ONGOING or RUNUP`.